Hardcode user scores category as "day"
